Séparation des fichiers. Ajout de la création de la classe générale et la classe des commandes
